major refactoring in smaller units for easier maintenance and overview - tests pass
