+++ ddoc/utils/index.js
@@ -1,0 +1,3 @@
+module.exports = {
+  loadFile: require('./load-file')
+}
