+++ ddoc/views/index.js
@@ -1,0 +1,7 @@
+module.exports = {
+  lib: require('./lib'),
+  list_by_type: require('./list-by-type'),
+  list_by_type_and_filter: require('./list-by-type-and-filter'), 
+  relations: require('./relations'),
+  relations_with_filters: require('./relations-with-filters')
+}
